backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
channels==4.0.0
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-nextjs==2.2.1
djangorestframework==3.14.0
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
